+++ nntrainer/tensor/short_tensor.cpp
@@ -470,11 +470,11 @@ void ShortTensor::copy(const void *buf) {
 }
 
 void ShortTensor::save_quantization_info(std::ostream &file) {
-  checkedWrite(file, (char *)&qscheme, sizeof(uint8_t),
+  checkedWrite(file, (char *)&qscheme, sizeof(uint16_t),
                "[ShortTensor::save] failed to write quantization information");
 }
 void ShortTensor::read_quantization_info(std::ifstream &file) {
-  checkedRead(file, (char *)&qscheme, sizeof(uint8_t),
+  checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
               "[ShortTensor::read] failed to read quantization information");
 }
